api_server/requirements.txt
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.7.0
cryptography==2.8
Flask==1.1.2
Flask-AuditLog==1.0
Flask-Cors==3.0.8
Flask-No-Cache==1.0
Flask-SSLify==0.1.5
Flask-Testing==0.7.1
google-api-core==1.20.0
google-auth==1.17.1
google-cloud-core==1.3.0
google-cloud-datastore==1.12.0
google-cloud-firestore==1.7.0
googleapis-common-protos==1.52.0
grpcio==1.29.0
idna==2.9
inflection==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
jwkaas==0.4
MarkupSafe==1.1.1
openapi-spec-validator==0.2.8
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.2
six==1.15.0
swagger-ui-bundle==0.0.6
urllib3==1.25.9
Werkzeug==1.0.1
